+++ src/main/java/com/lilittlecat/chatgpt/window/ChatGPTToolWindow.java
@@ -9,7 +9,6 @@ import com.intellij.ui.jcef.JBCefCookieManager;
 import com.lilittlecat.chatgpt.setting.ChatGPTSettingsState;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpEntity;
-import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
@@ -59,25 +58,27 @@ public class ChatGPTToolWindow extends SimpleToolWindowPanel {
             if (StringUtils.isNotBlank(sessionToken)) {
                 // check the token is right or not
 
-                String cookie = sessionTokenName + "=" + sessionToken;
-                HttpGet httpGet = new HttpGet("https://chat.openai.com/chat");
-                httpGet.setHeader("Connection", "keep-alive");
-                httpGet.setHeader("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.1 Safari/605.1.15");
-                httpGet.setHeader("Cookie", cookie);
-                httpGet.setHeader("Origin", "https://auth0.openai.com");
-                httpGet.setHeader("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8");
-                httpGet.setHeader("Accept-Language", "en-US,en;q=0.9");
-                try (CloseableHttpClient httpClient = HttpClients.createDefault();
-                     CloseableHttpResponse httpResponse = httpClient.execute(httpGet)) {
-                    HttpEntity httpEntity = httpResponse.getEntity();
-                    String bodyString = EntityUtils.toString(httpEntity);
-                    if (httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_OK && !bodyString.contains("Welcome to ChatGPT")) {
-                        JBCefCookie jbCefCookie = new JBCefCookie(sessionTokenName, sessionToken, "chat.openai.com", "/", true, true);
-                        jbCefCookieManager.setCookie("https://chat.openai.com", jbCefCookie);
-                    }
-                } catch (IOException e) {
-                    LOG.error("Error when check session token: ", e);
-                }
+                // 2022.12.16 with Cloudflare, can not check token by http get
+
+//                String cookie = sessionTokenName + "=" + sessionToken;
+//                HttpGet httpGet = new HttpGet("https://chat.openai.com/chat");
+//                httpGet.setHeader("Connection", "keep-alive");
+//                httpGet.setHeader("User-Agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.1 Safari/605.1.15");
+//                httpGet.setHeader("Cookie", cookie);
+//                httpGet.setHeader("Origin", "https://auth0.openai.com");
+//                httpGet.setHeader("Accept", "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8");
+//                httpGet.setHeader("Accept-Language", "en-US,en;q=0.9");
+//                try (CloseableHttpClient httpClient = HttpClients.createDefault();
+//                     CloseableHttpResponse httpResponse = httpClient.execute(httpGet)) {
+//                    HttpEntity httpEntity = httpResponse.getEntity();
+//                    String bodyString = EntityUtils.toString(httpEntity);
+//                    if (httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_OK && !bodyString.contains("Welcome to ChatGPT")) {
+                JBCefCookie jbCefCookie = new JBCefCookie(sessionTokenName, sessionToken, "chat.openai.com", "/", true, true);
+                jbCefCookieManager.setCookie("https://chat.openai.com", jbCefCookie, null);
+//                    }
+//                } catch (IOException e) {
+//                    LOG.error("Error when check session token: ", e);
+//                }
             }
         }
         // get session token after login, fill it in settings
@@ -86,10 +87,12 @@ public class ChatGPTToolWindow extends SimpleToolWindowPanel {
             String currentSessionToken = null;
             while (currentSessionToken == null) {
                 List<JBCefCookie> cookies = jbCefCookieManager.getCookies();
-                for (JBCefCookie cookie : cookies) {
-                    if (cookie.getName().equals(sessionTokenName)) {
-                        currentSessionToken = cookie.getValue();
-                        ChatGPTSettingsState.getInstance().update(currentSessionToken);
+                if (!cookies.isEmpty()) {
+                    for (JBCefCookie cookie : cookies) {
+                        if (cookie.getName().equals(sessionTokenName)) {
+                            currentSessionToken = cookie.getValue();
+                            ChatGPTSettingsState.getInstance().update(currentSessionToken);
+                        }
                     }
                 }
                 try {